chore: :truck: Update route for commercial module pages and components
