+++ pages/commercial/coupons/index.tsx
@@ -1,22 +1,22 @@
 import {
-  TDiscountPayload,
-  TNewDiscountPayload,
+  TCouponPayload,
+  TNewCouponPayload,
 } from '@/services/commercialApi/types';
 import {
-  createDiscount,
-  deleteDiscount,
-  getDiscounts,
-  updateDiscount,
-} from '@/services/commercialApi/discounts';
+  createCoupon,
+  deleteCoupon,
+  getCoupons,
+  updateCoupon,
+} from '@/services/commercialApi';
 
-import { Alert } from '@/components/alerts';
+import { Alert } from '@/components/commercialComponents/alerts';
 import { COMMERCIAL_CONSTANTS } from '@/constants/commercial';
-import { CreateDiscountModal } from '@/components/CreateModal';
-import { DiscountsTable } from '@/components/CommercialTable';
-import { EditDiscountModal } from '@/components/EditModal';
+import { CouponsTable } from '@/components/commercialComponents/CommercialTable';
+import { CreateCouponModal } from '@/components/commercialComponents/CreateModal';
+import { EditCouponModal } from '@/components/commercialComponents/EditModal';
 import Head from 'next/head';
 import Link from 'next/link';
-import { TDiscount } from '@/types';
+import { TCoupon } from '@/types';
 import { useDelete } from '@/hooks/useDelete';
 import { useFetch } from '@/hooks/useFetch';
 import { useState } from 'react';
@@ -24,13 +24,13 @@ import { useState } from 'react';
 const CITY_OPTIONS = COMMERCIAL_CONSTANTS.regions;
 const USER_TYPE_OPTIONS = COMMERCIAL_CONSTANTS.userTypes;
 
-const Discounts = () => {
-  const { data: discounts, setData: setDiscounts } = useFetch<TDiscount[]>(
-    getDiscounts,
+const Coupons = () => {
+  const { data: coupons, setData: setCoupons } = useFetch<TCoupon[]>(
+    getCoupons,
     []
   );
 
-  const [idDiscountToEdit, setIdDiscountToEdit] = useState('');
+  const [idCouponToEdit, setIdCouponToEdit] = useState('');
 
   const [openCreateModal, setOpenCreateModal] = useState(false);
   const [openEditModal, setOpenEditModal] = useState(false);
@@ -47,31 +47,35 @@ const Discounts = () => {
     if (formContext) {
       const formData = new FormData(formContext);
       const data = Object.fromEntries(formData.entries());
+
       const startDate = data.startDate.toString().replace('T', ' ') + ':00';
       const endDate = data.endDate.toString().replace('T', ' ') + ':00';
 
-      const newDiscount: TNewDiscountPayload = {
-        name: data.name as string,
-        endDate,
-        startDate,
-        description: data.description as string,
-        discountPercentage: parseInt(
-          (data.discountPercentage as string) || '0'
-        ),
-        maxDiscount: parseInt((data.maxDiscount as string) || '0'),
-        discountValue: parseInt((data.discountValue as string) || '0'),
-        minValue: parseInt((data.minValue as string) || '0'),
-        isActive: true,
-        city: data.city as string,
-        userType: data.userType as string,
+      const newCoupon: TNewCouponPayload = {
+        amountCreated: parseInt(data.amount as string),
+        strategy: {
+          name: data.name as string,
+          endDate,
+          startDate,
+          description: data.description as string,
+          discountPercentage: parseInt(
+            (data.discountPercentage as string) || '0'
+          ),
+          maxDiscount: parseInt((data.maxDiscount as string) || '0'),
+          discountValue: parseInt((data.discountValue as string) || '0'),
+          minValue: parseInt((data.minValue as string) || '0'),
+          city: data.city as string,
+          userType: data.userType as string,
+          isActive: true,
+        },
       };
 
       try {
-        const response = await createDiscount(newDiscount);
-        setDiscounts([...(discounts ?? []), response]);
+        const response = await createCoupon(newCoupon);
+        setCoupons([...(coupons ?? []), response]);
         setAlertOptions({
           severity: 'success',
-          message: 'Promoción creada exitosamente',
+          message: 'Cupón creado exitosamente',
         });
         setOpenCreateModal(false);
         setOpenAlert(true);
@@ -79,7 +83,7 @@ const Discounts = () => {
       } catch (err) {
         setAlertOptions({
           severity: 'error',
-          message: 'Error al crear la promoción',
+          message: 'Error al crear el cupón',
         });
         setOpenAlert(true);
         return false;
@@ -98,32 +102,30 @@ const Discounts = () => {
       const data = Object.fromEntries(formData.entries());
       const startDate = data.startDate.toString().replace('T', ' ') + ':00';
       const endDate = data.endDate.toString().replace('T', ' ') + ':00';
-
-      const editedDiscount = {
-        name: data.name as string,
-        endDate,
-        startDate,
-        description: data.description as string,
-        discountPercentage: parseInt(
-          (data.discountPercentage as string) || '0'
-        ),
-        maxDiscount: parseInt((data.maxDiscount as string) || '0'),
-        discountValue: parseInt((data.discountValue as string) || '0'),
-        minValue: parseInt((data.minValue as string) || '0'),
+      const editedCoupon = {
+        strategy: {
+          name: data.name as string,
+          endDate,
+          startDate,
+          description: data.description as string,
+          discountPercentage: parseInt(
+            (data.discountPercentage as string) || '0'
+          ),
+          maxDiscount: parseInt((data.maxDiscount as string) || '0'),
+          discountValue: parseInt((data.discountValue as string) || '0'),
+          minValue: parseInt((data.minValue as string) || '0'),
+        },
       };
 
       try {
-        const response = await updateDiscount(
-          id,
-          editedDiscount as TDiscountPayload
+        const response = await updateCoupon(id, editedCoupon as TCouponPayload);
+        const newCoupons = coupons?.map((coupon) =>
+          coupon.id === id ? response : coupon
         );
-        const newDiscounts = discounts?.map((discount) =>
-          discount.id === id ? response : discount
-        );
-        setDiscounts(newDiscounts ?? []);
+        setCoupons(newCoupons ?? []);
         setAlertOptions({
           severity: 'success',
-          message: 'Promoción editada exitosamente',
+          message: 'Cupón editado exitosamente',
         });
         setOpenEditModal(false);
         setOpenAlert(true);
@@ -131,36 +133,35 @@ const Discounts = () => {
       } catch (error) {
         setAlertOptions({
           severity: 'error',
-          message: 'Error al editar la promoción',
+          message: 'Error al editar el cupón',
         });
         setOpenAlert(true);
         return false;
       }
     }
-
     return false;
   };
 
-  const { setIdToDelete: setIdDiscountToDelete, setDeleteConfirm } =
-    useDelete<TDiscount>(
-      discounts,
-      setDiscounts,
+  const { setIdToDelete: setIdCouponToDelete, setDeleteConfirm } =
+    useDelete<TCoupon>(
+      coupons,
+      setCoupons,
       setAlertOptions,
       setOpenAlert,
-      'discount',
-      deleteDiscount
+      'coupon',
+      deleteCoupon
     );
 
   return (
     <>
       <Head>
-        <title>Promociones</title>
+        <title>Cupones</title>
         <meta name='description' content='Cupones' />
       </Head>
       <header>
         <nav className='flex justify-between items-center h-16 px-8 bg-blue-500 text-white'>
           <section className='brand'>
-            <Link href='./'>
+            <Link href='./coupons'>
               <h1 className='text-4xl font-bold'>SSMU</h1>
             </Link>
           </section>
@@ -184,16 +185,16 @@ const Discounts = () => {
         <aside className='2xl:col-span-1 md:col-span-2 p-8 bg-blue-50 '>
           <h2 className='text-xl font-semibold'>Marketing</h2>
           <ul className='py-4'>
+            <li className={`px-4 py-1 rounded-full font-semibold`}>
+              <Link href='../commercial/discounts'>&#8226; Promociones</Link>
+            </li>
             <li
               className={`px-4 py-1 rounded-full font-semibold bg-blue-500 text-white`}
             >
-              <Link href='../discounts'>&#8226; Promociones</Link>
+              <Link href='../commercial/coupons'>&#8226; Cupones</Link>
             </li>
             <li className={`px-4 py-1 rounded-full font-semibold`}>
-              <Link href='../coupons'>&#8226; Cupones</Link>
-            </li>
-            <li className={`px-4 py-1 rounded-full font-semibold`}>
-              <Link href='../coupons/validate'>&#8226; Validar</Link>
+              <Link href='../commercial/coupons/validate'>&#8226; Validar</Link>
             </li>
           </ul>
         </aside>
@@ -201,9 +202,7 @@ const Discounts = () => {
         <section className='col-span-5 mx-16 my-8 px-16 py-8 bg-white rounded-xl overflow-y-auto'>
           <header className='flex justify-between items-center'>
             <div>
-              <h1 className='text-4xl font-semibold text-blue-500'>
-                Promociones
-              </h1>
+              <h1 className='text-4xl font-semibold text-blue-500'>Cupones</h1>
             </div>
             <div>
               <button
@@ -212,41 +211,39 @@ const Discounts = () => {
                 }}
                 className='text-lg text-blue-500 border-4 border-blue-500 px-2 py-1 rounded-lg'
               >
-                Nueva promoción +
+                Nuevo Cupón +
               </button>
             </div>
           </header>
 
           <main className='my-8 min-h-[80%]'>
-            <DiscountsTable
-              elements={discounts ?? []}
-              setIdToEdit={setIdDiscountToEdit}
+            <CouponsTable
+              elements={coupons ?? []}
               setOpenEdit={setOpenEditModal}
-              setIdToDelete={setIdDiscountToDelete}
+              setIdToEdit={setIdCouponToEdit}
+              setIdToDelete={setIdCouponToDelete}
             />
           </main>
         </section>
 
-        <CreateDiscountModal
+        <CreateCouponModal
           open={openCreateModal}
           setOpen={setOpenCreateModal}
           handleCreate={handleCreate}
           regionOptions={CITY_OPTIONS}
           userTypeOptions={USER_TYPE_OPTIONS}
         />
-        {idDiscountToEdit && (
-          <EditDiscountModal
+        {idCouponToEdit && (
+          <EditCouponModal
             open={openEditModal}
             setOpen={setOpenEditModal}
-            setIdToEdit={setIdDiscountToEdit}
+            setIdToEdit={setIdCouponToEdit}
+            data={
+              coupons?.find((coupon) => coupon.id === idCouponToEdit) as TCoupon
+            }
             handleEdit={handleEdit}
             regionOptions={CITY_OPTIONS}
             userTypeOptions={USER_TYPE_OPTIONS}
-            data={
-              discounts?.find(
-                (discount) => discount.id === idDiscountToEdit
-              ) as TDiscount
-            }
           />
         )}
         <Alert
@@ -262,4 +259,4 @@ const Discounts = () => {
   );
 };
 
-export default Discounts;
+export default Coupons;
